+++ src/main/java/com/example/demo/dao/KospiInfo.java
@@ -65,37 +65,4 @@ public class KospiInfo extends CommonSearch {
         }
         return stocks;
     }
-
-    public boolean checkDb(WebElement element) {
-        if (stockRepository.findAll().contains(stockRepository.findByName(getTitle(element))))
-            return true;
-        return false;
-    }
-
-    public Stock makeStock(int i, int j) {
-        WebElement element = getDriver().findElement(By.xpath("//*[@id=\"wrap\"]/div[" + i + "]/div/div[3]/dl[" + j + "]"));
-        if (j % 10 == 0)
-            logger.info("{}part {}번째 data, title : {}",i, j, getTitle(element));
-
-        if (checkDb(element))
-            return stockRepository.findByName(getTitle(element)).realDataUpdate(getTitle(element), getInfo(element).get(1), getInfo(element).get(2), getInfo(element).get(3), getUrl(element));
-        return new Stock(getTitle(element), getInfo(element).get(1), getInfo(element).get(2), getInfo(element).get(3), getUrl(element));
-    }
-
-    public List<WebElement> getChilds(WebElement element) {
-        return element.findElements(By.xpath(".//*"));
-    }
-
-    public String getUrl(WebElement element) {
-        String url = getChilds(element).get(0).getAttribute("onclick");
-        return url.substring(8, url.length() -2);
-    }
-
-    public String getTitle(WebElement element) {
-        return getChilds(element).get(0).getAttribute("title");
-    }
-
-    public List<String> getInfo(WebElement element) {
-        return Arrays.asList(element.getText().split("\n"));
-    }
 }
